refactor: extract snapshot collector functions into lib/snapshot-collector.ts (Phase 3f)
- Created createSnapshotCollector(deps) DI factory with 5 functions
- Extracted: getApiSnapshotCount, updateApiSnapshot, installApiHook,
  normalizeSnapshotCollectOptions, collectSnapshotPayload
- Adapted document→doc in factory versions
- Adapted job51DetailBackfillRunId/Chain → pipelineState getter/setter
- Restored window.addEventListener("message", ...) hook backed by factory fn
- content.ts reduced by ~378L
